Restructure/re-route component directory
