+++ app/containers/VideoSetPage.js
@@ -1,4 +1,4 @@
 import { connect } from 'react-redux';
-import VideoSet from '../components/VideoSet';
+import VideoSet from '../components/VideoComponents/VideoSet';
 
 export default connect()(VideoSet);
